기본 틀2

+++ src/main/java/com/github/sc_first_project/web/repository/likes/LikesRepository.java
@@ -2,5 +2,5 @@ package com.github.sc_first_project.web.repository.likes;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LikesRepository extends JpaRepository<LikesEntity, Integer> {
+public interface LikesRepository extends JpaRepository<LikesEntity, Long> {
 }
